modified database name to SA46Team05BESNETProject/without spaces. added two crystal reports .rpt and a form for crystal report viewer

+++ SA46Team05BESNETProject/Member.cs
@@ -24,7 +24,7 @@ namespace SA46Team05BESNETProject
         public string MemberName { get; set; }
         public string Phone { get; set; }
         public string Gender { get; set; }
-        public Nullable<double> Age { get; set; }
+        public Nullable<int> Age { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Transaction> Transactions { get; set; }
